Add factory for meta models needed by graphene
Graphene django requires a class Meta that specifies model and fields for a node.

Since we plan to generate this part dynamically we need a factory to do that for us.

Since the model is fairly simple and we don't really need to change anything,
we can have a plain class factory and don't need to use type or derivatives.